+++ src/datapack/mod.rs
@@ -8,7 +8,7 @@ pub use command::{Command, Condition, Execute};
 pub use function::Function;
 pub use namespace::Namespace;
 
-use std::{collections::HashMap, ops::RangeInclusive, sync::Mutex};
+use std::{borrow::Cow, collections::BTreeMap, ops::RangeInclusive, sync::Mutex};
 
 use crate::{
     util::compile::{CompileOptions, CompilerState, MutCompilerState},
@@ -23,21 +23,28 @@ pub struct Datapack {
     description: String,
     pack_format: u8,
     supported_formats: Option<RangeInclusive<u8>>,
-    namespaces: HashMap<String, Namespace>,
+    main_namespace_name: String,
+    namespaces: BTreeMap<String, Namespace>,
+    /// Scoreboard name -> (criteria, display name)
+    scoreboards: BTreeMap<String, (Option<String>, Option<String>)>,
+    uninstall_commands: Vec<Command>,
     custom_files: VFolder,
 }
 
 impl Datapack {
-    pub const LATEST_FORMAT: u8 = 48;
+    pub const LATEST_FORMAT: u8 = 61;
 
     /// Create a new Minecraft datapack.
     #[must_use]
-    pub fn new(pack_format: u8) -> Self {
+    pub fn new(main_namespace_name: impl Into<String>, pack_format: u8) -> Self {
         Self {
             description: String::from("A Minecraft datapack created with shulkerbox"),
             pack_format,
             supported_formats: None,
-            namespaces: HashMap::new(),
+            main_namespace_name: main_namespace_name.into(),
+            namespaces: BTreeMap::new(),
+            scoreboards: BTreeMap::new(),
+            uninstall_commands: Vec::new(),
             custom_files: VFolder::new(),
         }
     }
@@ -65,6 +72,7 @@ impl Datapack {
     /// # Errors
     /// - If loading the directory fails
     #[cfg(feature = "fs_access")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "fs_access")))]
     pub fn with_template_folder<P>(self, path: P) -> std::io::Result<Self>
     where
         P: AsRef<std::path::Path>,
@@ -90,24 +98,49 @@ impl Datapack {
     }
 
     /// Mutably get a namespace by name or create a new one if it doesn't exist.
-    pub fn namespace_mut(&mut self, name: &str) -> &mut Namespace {
+    pub fn namespace_mut(&mut self, name: impl Into<String>) -> &mut Namespace {
+        let name = name.into();
         self.namespaces
-            .entry(name.to_string())
+            .entry(name.clone())
             .or_insert_with(|| Namespace::new(name))
     }
 
     /// Add a function to the tick function list.
-    pub fn add_tick(&mut self, function: &str) {
+    pub fn add_tick(&mut self, function: impl Into<String>) {
         self.namespace_mut("minecraft")
             .tag_mut("tick", tag::TagType::Function)
-            .add_value(tag::TagValue::Simple(function.to_string()));
+            .add_value(tag::TagValue::Simple(function.into()));
     }
 
     /// Add a function to the load function list.
-    pub fn add_load(&mut self, function: &str) {
+    pub fn add_load(&mut self, function: impl Into<String>) {
         self.namespace_mut("minecraft")
             .tag_mut("load", tag::TagType::Function)
-            .add_value(tag::TagValue::Simple(function.to_string()));
+            .add_value(tag::TagValue::Simple(function.into()));
+    }
+
+    /// Register a scoreboard.
+    pub fn register_scoreboard(
+        &mut self,
+        name: impl Into<String>,
+        criteria: Option<impl Into<String>>,
+        display_name: Option<impl Into<String>>,
+    ) {
+        self.scoreboards.insert(
+            name.into(),
+            (criteria.map(Into::into), display_name.map(Into::into)),
+        );
+    }
+
+    /// Scoreboards registered in the datapack.
+    #[must_use]
+    pub fn scoreboards(&self) -> &BTreeMap<String, (Option<String>, Option<String>)> {
+        &self.scoreboards
+    }
+
+    /// Add commands to the uninstall function.
+    pub fn add_uninstall_commands(&mut self, commands: Vec<Command>) {
+        self.uninstall_commands.extend(commands);
     }
 
     /// Add a custom file to the datapack.
@@ -132,8 +165,71 @@ impl Datapack {
         root_folder.add_file("pack.mcmeta", mcmeta);
         let mut data_folder = VFolder::new();
 
+        let mut modified_namespaces = self
+            .namespaces
+            .iter()
+            .map(|(name, namespace)| (name.as_str(), Cow::Borrowed(namespace)))
+            .collect::<BTreeMap<_, _>>();
+
+        let mut uninstall_commands = options
+            .uninstall_function
+            .then_some(Cow::Borrowed(&self.uninstall_commands));
+
+        if !self.scoreboards.is_empty() {
+            let main_namespace = modified_namespaces
+                .entry(&self.main_namespace_name)
+                .or_insert_with(|| Cow::Owned(Namespace::new(&self.main_namespace_name)));
+            let register_scoreboard_function = main_namespace
+                .to_mut()
+                .function_mut("sb/register_scoreboards");
+            for (name, (criteria, display_name)) in &self.scoreboards {
+                let mut creation_command = format!(
+                    "scoreboard objectives add {name} {criteria}",
+                    criteria = criteria.as_deref().unwrap_or("dummy")
+                );
+                if let Some(display_name) = display_name {
+                    creation_command.push(' ');
+                    creation_command.push_str(display_name);
+                }
+                register_scoreboard_function.add_command(Command::Raw(creation_command));
+
+                if let Some(uninstall_commands) = uninstall_commands.as_mut() {
+                    uninstall_commands
+                        .to_mut()
+                        .push(Command::Raw(format!("scoreboard objectives remove {name}")));
+                }
+            }
+
+            let minecraft_namespace = modified_namespaces
+                .entry("minecraft")
+                .or_insert_with(|| Cow::Owned(Namespace::new("minecraft")));
+            minecraft_namespace
+                .to_mut()
+                .tag_mut("load", tag::TagType::Function)
+                .values_mut()
+                .insert(
+                    0,
+                    tag::TagValue::Simple(format!(
+                        "{}:sb/register_scoreboards",
+                        self.main_namespace_name
+                    )),
+                );
+        }
+
+        if let Some(uninstall_commands) = uninstall_commands {
+            if !uninstall_commands.is_empty() {
+                let main_namespace = modified_namespaces
+                    .entry(&self.main_namespace_name)
+                    .or_insert_with(|| Cow::Owned(Namespace::new(&self.main_namespace_name)));
+                let uninstall_function = main_namespace.to_mut().function_mut("uninstall");
+                uninstall_function
+                    .get_commands_mut()
+                    .extend(uninstall_commands.into_owned());
+            }
+        }
+
         // Compile namespaces
-        for (name, namespace) in &self.namespaces {
+        for (name, namespace) in modified_namespaces {
             let namespace_folder = namespace.compile(&options, &compiler_state);
             data_folder.add_existing_folder(name, namespace_folder);
         }
@@ -180,7 +276,7 @@ mod tests {
     fn test_datapack() {
         let template_dir = tempfile::tempdir().expect("error creating tempdir");
 
-        let mut dp = Datapack::new(Datapack::LATEST_FORMAT)
+        let mut dp = Datapack::new("main", Datapack::LATEST_FORMAT)
             .with_description("My datapack")
             .with_template_folder(template_dir.path())
             .expect("error reading template folder");
@@ -193,7 +289,7 @@ mod tests {
 
     #[test]
     fn test_generate_mcmeta() {
-        let dp = &Datapack::new(Datapack::LATEST_FORMAT).with_description("foo");
+        let dp = &Datapack::new("main", Datapack::LATEST_FORMAT).with_description("foo");
         let state = Mutex::new(CompilerState::default());
         let mcmeta = generate_mcmeta(dp, &CompileOptions::default(), &state);