+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - support for commands using macros
+- support for registering scoreboards (automatic creation and deletion)
 
 ### Changed
 - use "return" command for conditionals instead of data storage when using supported pack format
+- update latest datapack format to 61
 
 ### Removed